Fix lint errors. Ignore blank lines

+++ sync_on_save.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import print_function  # for Python 2.6/SublimeText 2 compatibility
+from __future__ import print_function  # for Python 2.x/SublimeText 2
 import sublime_plugin
 import sublime
 
@@ -30,6 +30,7 @@ EXCLUDE_PATTERNS = [
     '.DS_Store',
 ] + TRIGGER_FILES
 
+
 def debug(message):
     print("SyncOnSave:", message, file=sys.stderr)
 
@@ -37,16 +38,18 @@ def debug(message):
 class SyncOnSave(sublime_plugin.EventListener):
     """Sync changes to remote machine/directory on save
 
-    Uses rsync to sync to a remote machine. To specify what to sync where, put a
-    .syncsettings trigger file in the directory you want to sync.
-    It must be the directory or some parent directory of your current working file.
-    Upon saving any file within the directory with the trigger file,
-    the directory containing the trigger file will be synced to all
-    destinations listed in the trigger file as per rsync syntax e.g.
+    Uses rsync to sync to a remote machine. To specify what to sync where,
+    put a ".syncsettings" trigger file in the directory you want to sync.
+    It must be the directory or some parent directory of your current
+    working file. Upon saving any file within the directory with the
+    trigger file, the directory containing the trigger file will be
+    synced to all destinations listed in the trigger file as per
+    rsync syntax e.g:
 
-    my.remote.machine:/home/freider/proj
-    192.168.0.12:/var/www
-    /local/dir
+        my.remote.machine:/home/freider/proj
+        192.168.0.12:/var/www
+        /local/dir
+
     """
     def sync_working_dir(self, view):
         test_dir = view.file_name()
@@ -67,13 +70,22 @@ class SyncOnSave(sublime_plugin.EventListener):
         with open(trigger_file) as f:
             for line in f:
                 remote_location = line.strip()
-                view.set_status("SyncOnSave", "Uploading to {0}".format(remote_location))
+                if not remote_location:
+                    continue
+                view.set_status(
+                    "SyncOnSave",
+                    "Uploading to {0}".format(remote_location)
+                )
                 cmd = ['rsync', '-av', '--delete']
                 for pattern in EXCLUDE_PATTERNS:
                     cmd += ['--exclude', pattern]
                 cmd += [local_dir, remote_location]
                 print(cmd)
-                p = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+                p = subprocess.Popen(
+                    cmd,
+                    stdout=subprocess.PIPE,
+                    stderr=subprocess.STDOUT
+                )
                 for line in p.stdout:
                     debug(line.decode('utf8').rstrip('\n'))
                 status = p.wait()
